requirements.txt
Django==1.8.14
pytz==2016.6.1
psycopg2==2.6.2
djangorestframework==3.4.1
Markdown==2.6.6
django-filter==0.13.0
django-extensions==1.6.7
django-autocomplete-light==3.1.6
django-modeltranslation==0.11